Move PondWater module from Model namespace to Model.Aqua namespace. Also add some <summary> tags to remove warnings from VisualStudio
