+++ Models/Aqua/WaterPropertiesData.cs
@@ -8,19 +8,52 @@ namespace Models.Aqua
 
     /// <summary>
     /// Stores the information about the water but not the volume of it.
+    /// Applies to any given volume/amount of this water.
     /// </summary>
     [Serializable]
     public class WaterProperties
         {
 
-        public double Temperature; 
+        /// <summary>
+        /// Temperature of the water (oC)
+        /// </summary>
+        public double Temperature;
+ 
+        /// <summary>
+        /// Salinity (kg/m^3)
+        /// </summary>
         public double Salinity; 
+        
+        /// <summary>
+        /// PH 
+        /// </summary>
         public double PH;
+
+        /// <summary>
+        /// Nitrogen (kg/m^3)
+        /// </summary>
         public double N;
+
+        /// <summary>
+        /// Phosphorus (kg/m^3)
+        /// </summary>
         public double P;
+
+        /// <summary>
+        /// Total Suspended Solids (kg/m^3)
+        /// </summary>
         public double TSS;
 
 
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="Temperature">Temperature (oC)</param>
+        /// <param name="Salinity">Salinity (kg/m^3)</param>
+        /// <param name="PH">PH</param>
+        /// <param name="N">Nitrogen (kg/m^3)</param>
+        /// <param name="P">Phosporus (kg/m^3)</param>
+        /// <param name="TSS">Total Suspended Solids (kg/m^3)</param>
         public WaterProperties(double Temperature, double Salinity, double PH, double N, double P, double TSS)
             {
             this.Temperature = Temperature;
@@ -31,7 +64,9 @@ namespace Models.Aqua
             this.TSS = TSS;
             }
 
-
+        /// <summary>
+        /// Zero all the water properties
+        /// </summary>
         public void ZeroProperties()
             {
             this.Temperature = 0;
